Improve CygWin instructions
I just clean-installed the computer and had to rebuild my CygWin environment to generate the last couple days of drafts.  I'd like to make sure the info I needed to get back to a workable state is in the docs, so I don't have to re-research the next time I flatten the machine.

Oh, and it might help someone else, too.  ;-)