Check for origin/master existing before setup
Not having origin/master present means that GitHub will fixate on the gh-pages
branch.  That's not good for this setup and I have found that circle CI in
particular gets its panties in a twist in such a way that you can't easily
recover.  This checks for origin/master existing and asks the user to push if
it isn't.