Don't update issues/pulls too often
GitHub implements rate control for its API.  We can make at most 60 requests
every hour without tripping this limit.  There is a higher limit for
authenticated requests, but authenticating in CI is hard enough for pushing.

Thus, this limits the the rate at which we hit the API.  Only update once per
hour.  Copy the latest copy of the issues.json and pulls.json file from the
gh-issues branch rather than download the latest within this time.  This will
be detected as a no-op and the gh-issues branch won't be updated.