Be a little more tolerant of old tags
The `git describe` command by default ignores lightweight tags, but
this results in generating targets that are incorrect.  Rather than
insist that people correct their repos, just accept the old tags as
they are and make things work.  This change won't cause lightweight
tags to be accepted when uploading to datatracker, it will only fix
builds of older versions of drafts.