Make the HTML narrower
This doesn't quite get down to 90 characters per line.  I got most of the way through the "c" on the fourth alphabet, so it's not ideal, but it's still an improvement over previous.  On Mac, Helvetica Neue is slightly wider, so it only gets a partial pixel of the "c", so all win there.

A narrower page allows the TOC to be shown on the side earlier.  It could be made bigger instead, but I think that the size we have is OK.

@mnot, WDYT?