A bunch of fixes
I couldn't be bothered splitting these up.

default-branch.py looks locally for a branch name and uses that, either
in origin/HEAD (or $GIT_REMOTE/HEAD) or just HEAD.

It also requires a token if it is going to use the API.

The token variables are rationalized a little.  GITHUB_PUSH_TOKEN is
what is used for pushing.  GITHUB_API_TOKEN is what is used for things
that need the API.

id.mk uses the token to get the default branch, it then exports that for
use by scripts.

I've also updated some of the documentation.