Page numbers are the past
This causes drafts to be rendered in the same way as RFCs, by default.
It opens up the possibility of using tools *other than* rfcdiff for
examining changes between versions.  As textual output is often easier
to handle than HTML when it comes to diff tools, this is a strict
improvement.

This will mean that submissions look different, if the text versions are
submitted.  That is unlikely.  What is more likely is that the submitted
version will be paginated, as that is the xml2rfc default.  This means
that diffs for editors copy (and branches) with the submitted versions
will look worse.  rfcdiff can handle that well enough; other tools
won't be any more usable for that case than they were before.

Overall, this seems to be a strict win.  There might be complaints (I
remain impressed and horrified, but unsurprised, at the reaction to the
lack of page numbers in published RFCs).  That's fine.  I believe that
most people using this toolchain use the HTML rendering primarily, as
that remains more accessible than text.

I am NOT adding a configuration option to disable this, but might be
convinced to.

Closes #251.