Fix a11y for pilcrows
This makes them visible always, but not emphasized so that they don't
distract sighted readers.  This means that screen readers will have this
bump always, but the feedback I received indicated that this was
something that people would prefer to have access to, even if it adds a
little extra noise.