Update @mnot's patch to make it more robust
This passes in source files rather than relying on the cwd being
correct.  It also only passes one file in to the extract metadata
script.  That file is based on the input.

For HTTP, it also caches what it learns to avoid the extra work.  This
might run afoul of the ancient bash on mac not liking it.  We'll see.