Run cleanup before making index files
In the case where `make` executes prerequisites in a different order
than they are specified, the index files can end up referencing content
that has been removed.  This avoids that problem, at a cost of running a
`.PHONY` target as a prerequisite of a file target (a pattern I don't
like).  Still, it's better than not having proper dependencies.

Closes #302.