Use the full tag or branch name
The way that GitHub Actions present the current ref is good, but my code
for getting the bare name was a little sloppy.

I've included code for making a "branch" name from a tag name.  That has
some problems, but this is merely annoying, not serious.  It might
create collisions in names, which will cause the contents of the tag to
compete with the contents of the branch.  Cleanup of the tag will occur
as soon as possible (default: 30 days).

Closes #307.