Fix GitHub pages index when using mmark.
Building of the index.html file for GitHub Pages is broken with the
`mmark` tool due to the metadata being in TOML rather than YAML format.

This results in an index being generated without any extracted metadata.

This change adds the TOML tool and differentiates by the header
characters ('%%%' vs '---') which format to parse.