Override install locations in CI
This should help with caching, which happens outside of the docker
image.  As a result, we need to use relative paths for everything AND
ensure that anything we might want to cache is kept in the working
directory directly.  Not in /tmp, which goes away, and not under /lib,
which does not exist when the cache is restored.